// src/presentation/features/core/pages/edi-purchase-order/types/TargetJson850.ts
export interface Document {
  code: string;
  output: Output;
}

export interface Output {
  interchanges: Interchange[];
  __version: string;
}

export interface Interchange {
  interchange_control_header_ISA: { [key: string]: string };
  groups: Group[];
  interchange_control_trailer_IEA: InterchangeControlTrailerIEA;
  delimiters: Delimiters;
}
// ...
export interface TransactionSet {
  heading: Heading;
  detail: Detail;
  summary: Summary;
  set: string;
}

export interface Detail {
  baseline_item_data_PO1_loop: BaselineItemDataPO1Loop[];
}

export interface BaselineItemDataPO1Loop {
  baseline_item_data_PO1: {
    assigned_identification_01: string;
    quantity_ordered_02: string;
    unit_or_basis_for_measurement_code_03: string;
    unit_price_04: string;
    basis_of_unit_price_code_05: string;
    product_service_id_qualifier_06: string;
    product_service_id_07: string;
    product_service_id_qualifier_08: string;
    product_service_id_09: string;
    product_service_id_qualifier_10: string;
    product_service_id_11: string;
    product_service_id_qualifier_12: string;
    product_service_id_13: string;
  };
  pricing_information_CTP_loop: PricingInformationCTPLoop[];
  product_item_description_PID_loop: ProductItemDescriptionPIDLoop[];
  item_physical_details_PO4: ItemPhysicalDetailsPO4[];
  line_item_schedule_SCH_loop: LineItemScheduleSCHLoop[];
}
// ...
export interface Heading {
  transaction_set_header_ST: TransactionSetHeaderST;
  beginning_segment_for_purchase_order_BEG: BeginningSegmentForPurchaseOrderBEG;
  reference_identification_REF: ReferenceIdentificationREF[];
  fob_related_instructions_FOB: FobRelatedInstructionsFOB[];
  sales_requirements_CSH: SalesRequirementsCSH[];
  terms_of_sale_deferred_terms_of_sale_ITD: TermsOfSaleDeferredTermsOfSaleITD[];
  date_time_reference_DTM: DateTimeReferenceDTM[];
  carrier_details_routing_sequence_transit_time_TD5: CarrierDetailsRoutingSequenceTransitTimeTD5[];
  name_N1_loop: NameN1Loop[];
}

export interface BeginningSegmentForPurchaseOrderBEG {
  transaction_set_purpose_code_01: string;
  purchase_order_type_code_02: string;
  purchase_order_number_03: string;
  date_05: string;
  contract_number_06: string;
}

export interface CarrierDetailsRoutingSequenceTransitTimeTD5 {
  routing_sequence_code_01: string;
  identification_code_qualifier_02: string;
  identification_code_03: string;
}

export interface DateTimeReferenceDTM {
  date_time_qualifier_01: string;
  date_02: string;
}

export interface FobRelatedInstructionsFOB {
  shipment_method_of_payment_01: string;
  location_qualifier_02: string;
  description_03: string;
}

export interface NameN1Loop {
  name_N1: NameN1;
  address_information_N3: AddressInformationN3[];
  geographic_location_N4: GeographicLocationN4[];
}

export interface AddressInformationN3 {
  address_information_01: string;
}

export interface GeographicLocationN4 {
  city_name_01: string;
  state_or_province_code_02: string;
  postal_code_03: string;
  country_code_04: string;
}

export interface NameN1 {
  entity_identifier_code_01: string;
  name_02: string;
  identification_code_qualifier_03: string;
  identification_code_04: string;
}

export interface ReferenceIdentificationREF {
  reference_identification_qualifier_01: string;
  reference_identification_02: string;
  description_03?: string;
}

export interface SalesRequirementsCSH {
  sales_requirement_code_01: string;
}

export interface TermsOfSaleDeferredTermsOfSaleITD {
  terms_type_code_01: string;
  terms_basis_date_code_02: string;
  terms_discount_percent_03: string;
  terms_discount_days_due_05: string;
}

export interface TransactionSetHeaderST {
  transaction_set_identifier_code_01: string;
  transaction_set_control_number_02: string;
}

export interface Summary {
  transaction_totals_CTT_loop: TransactionTotalsCTTLoop[];
  transaction_set_trailer_SE: TransactionSetTrailerSE;
}

export interface TransactionSetTrailerSE {
  number_of_included_segments_01: string;
  transaction_set_control_number_02: string;
}

export interface TransactionTotalsCTTLoop {
  transaction_totals_CTT: TransactionTotalsCTT;
}